part1 ok pelo war machine

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: mpasturi <mpasturi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/26 17:56:39 by mpasturi          #+#    #+#             */
-/*   Updated: 2020/02/29 19:49:00 by mpasturi         ###   ########.fr       */
+/*   Updated: 2020/02/29 21:16:11 by mpasturi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ void			*ft_memmove(void *dst, const void *src, size_t len);
 void			*ft_memset(void *b, int c, size_t len);
 int				ft_strlen(const char *s);
 int				ft_strlcpy(char *dst, const char *src, size_t dstsize);
-int				ft_strlcat(char *dst, const char *src, size_t dstsize);
+size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
 char			*ft_strchr(const char *s, int c);
 char			*ft_strrchr(const char *s, int c);
 char			*ft_strnstr(const char *haystack,
